# T1 & DWT multithreading decoding optimizations