This PR builds upon https://github.com/uclouvain/openjpeg/pull/783 and adds multithreading decoding optimizations. (The multithreading decoding optimizations are independant from the improvements of #783 but as they touch both t1.c , it was more convenient to build upon)

The main components of this PR are :
- Adds a threading API in thread.h/thread.c that adds an abstraction of POSIX mutex, conditions + joinable threads + high level  thread pool API. Implementations for pthread on Unix systems, Win32 API or stub if no OS threading detected, or if thread use explicitly disabled with CMake -DUSE_THREAD=OFF
- Add a opj_codec_set_threads() API & equivalent logic through the OPJ_NUM_THREADS=int/ALL_CPUS environment variable (so as to ease testing of existing code without modifying it to use opj_codec_set_threads())
- Use thread pool for parallel decoding of code blocks in C1
- Use thread pool for parallel processing of horizontal and vertical 1D DWT
- Add a -threads int/ALL_CPUS parameter to opj_decode
- Modify time measurement (on Linux) for opj_decode to avoid reporting the total CPU time
- Add a new .travis.yml target that runs with OPJ_NUM_THREADS=2

Benchmarking:

This has been tested with the following files :
C1: issue135.j2k (fom openjpeg-data, code blocks 32x32)
C2: Bretagne2.j2k (fom openjpeg-data, code blocks 32x32)
C3: 20160307_125117_0c74.jp2 (non public test file, 3 bands, 12 bits, 6600x2200 for band 1, 3300x2200 for bands 2 and 3, code blocks 64x64)
C4: issue135_vsc.jp2 ( issue135.j2k recoded by opj_compress -M 8, code blocks 64x64)
C5: issue135_raw.jp2 ( issue135.j2k recoded by opj_compress -M 1, code blocks 64x64)
C6: S2A_OPER_MSI_L1C_TL_MTI__20150819T171650_A000763_T30SWE_B05.jp2 (Sentinel 2 tile, 5490x5490, 1 band, 12 bits, code blocks 64x64)

Builds done with -DCMAKE_BUILD_TYPE=Release. Times measured are the smallest time of 2 consecutive runs reported by "OPJ_NUM_THREADS=4 opj_decompress -i $(INPUT_FILE) -o /tmp/out.ppm" in the "decode time: XXX ms" line

Machine & OS spec: Intel(R) Core(TM) i5 CPU 750 @ 2.67GHz, Linux 64 bit

Before = state of PR  #783
After = state of this PR

| compiler | C1 before (ms) | C1 after (ms) | delta % | C2 before (ms) | C2 after (ms) | delta % | C3 before (ms) | C3 after (ms) | delta % | C4 before (ms) | C4 after (ms) | delta % | C5 before (ms) | C5 after (ms) | delta % | C6 before (ms) | C6 after (ms) | delta % |
| --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- |
| GCC 4.4 | 730 | 359 | -50.8 | 1569 | 658 | -58.1 | 3920 | 1310 | -66.6 | 2060 | 819 | -60.2 | 1569 | 606 | -61.4 | 3770 | 1398 | -62.9 |
| GCC 4.6.2 | 710 | 343 | -51.7 | 1490 | 657 | -55.9 | 3860 | 1585 | -58.9 | 2060 | 793 | -61.5 | 1480 | 587 | -60.3 | 3710 | 1351 | -63.6 |
| GCC 4.8.0 | 740 | 376 | -49.2 | 1560 | 661 | -57.6 | 4040 | 1390 | -65.6 | 2130 | 763 | -64.2 | 1540 | 623 | -59.5 | 3910 | 1408 | -64.0 |
| GCC 5.2.0 | 730 | 379 | -48.1 | 1660 | 712 | -57.1 | 4080 | 1548 | -62.1 | 2050 | 700 | -65.9 | 1569 | 539 | -65.6 | 3890 | 1502 | -61.4 |
| GCC 5.3.0 | 720 | 382 | -46.9 | 1580 | 658 | -58.4 | 4000 | 1407 | -64.8 | 2040 | 814 | -60.1 | 1560 | 598 | -61.7 | 3830 | 1484 | -61.3 |
| CLang 3.7.0 | 740 | 430 | -41.9 | 1560 | 702 | -55.0 | 4080 | 1570 | -61.5 | 2420 | 880 | -63.6 | 1549 | 636 | -58.9 | 3929 | 1413 | -64.0 |

This work has been funded by Planet Labs