@mayeut 2016-05-26
Lots of good things, unfortunately, I don't have mush time to review this completely.
I think using native threads rather than OpenMP is a good thing as there will be more freedom for configuration in complex integrations.