@mayeut 2016-07-14
@detonin, it would probably be a good time to merge this PR or  #783 
Any idea ?