@detonin 2016-07-22
@rouault Thanks for this great work. 2 questions:
- Could you add the allocation checks suggested by @mayeut ?
- Does this multithreading implementation work on all platforms or only on Linux ? (didn't have the chance yet to test it on windows or macos and saw the conf you added in travis is linux)