@rouault 2016-08-11
I think all review comments have now been addressed