@rouault 2016-09-08
@mayeut  @detonin Anything preventing this PR from being merged ? This branch has been sync'ed today with master to fix merge conflicts, all CI tests pass and all comments from review have been addressed AFAICS.