@detonin 2016-09-08
@rouault Thanks for having kept the PR in sync and mergeable, and for having addressed the comments. From my point of view, it can now be merged indeed, especially as it does not break API nor ABI. 
@mayeut ok for you ?