@mayeut 2016-09-08
@detonin, is it possible to release v2.1.2 before merging this in something that would become v2.2.0 ?