@malaterre 2016-09-09
@mayeut I was about to say something like this ! However the root issue is still to have some kind of release branch where all CVE are fixed, which would simplify my life as a Debian package maintainer. And at the same time have git/master be the -next branch...
I did not post originally this comment in the bug report, fearing this would delay @rouault PR even more... (sorry Even)