@mayeut 2016-09-11
@malaterre, what could be done is to merge current master branch in openjpeg-2.1 branch (https://github.com/uclouvain/openjpeg/compare/openjpeg-2.1...master) while waiting for a release then merge this PR in master.

@detonin, is that ok ?