@detonin 2016-09-13
@malaterre thanks Mathieu for the merge, we'll proceed with the PR then