@detonin 2016-09-13
@rouault Many thanks Even for these great optimizations that are now merged in master and will be released soon (as v2.2.0).

I guess some of the still open PR will need to be updated before being able to merge them.